Switch to Geometry provider
