+++ faker/tests/__init__.py
@@ -163,13 +163,13 @@ class UtilsTestCase(unittest.TestCase):
             'faker.providers.currency',
             'faker.providers.date_time',
             'faker.providers.file',
+            'faker.providers.geometry',
             'faker.providers.internet',
             'faker.providers.job',
             'faker.providers.lorem',
             'faker.providers.misc',
             'faker.providers.person',
             'faker.providers.phone_number',
-            'faker.providers.point',
             'faker.providers.profile',
             'faker.providers.python',
             'faker.providers.ssn',
